+++ pkg/slices/bytes.go
@@ -1,5 +1,9 @@
 package slices
 
+import (
+	"bytes"
+)
+
 // BytesToStrings converts a slice of []byte into a slice of strings.
 func BytesToStrings(a [][]byte) []string {
 	s := make([]string, 0, len(a))
@@ -34,4 +38,29 @@ func CopyChunkedByteSlices(src [][]byte, chunkSize int) [][]byte {
 	}
 
 	return dst
+}
+
+// CompareSlice returns an integer comparing two slices of byte slices
+// lexicographically.
+// The result will be 0 if a==b, -1 if a < b, and +1 if a > b.
+func CompareSlice(a, b [][]byte) int {
+	i := 0
+	for i < len(a) && i < len(b) {
+		if v := bytes.Compare(a[i], b[i]); v == 0 {
+			i++
+			continue
+		} else {
+			return v
+		}
+	}
+
+	if i < len(b) {
+		// b is longer, so assume a is less
+		return -1
+	} else if i < len(a) {
+		// a is longer, so assume b is less
+		return 1
+	} else {
+		return 0
+	}
 }
