+++ pkg/slices/bytes_test.go
@@ -1,10 +1,15 @@
 package slices
 
 import (
+	"bytes"
+	"fmt"
 	"math"
 	"reflect"
 	"testing"
 	"unsafe"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/influxdata/platform/pkg/bytesutil"
 )
 
 func TestCopyChunkedByteSlices_oneChunk(t *testing.T) {
@@ -74,5 +79,68 @@ func TestCopyChunkedByteSlices_multipleChunks(t *testing.T) {
 	dst[0][5] = 'z'
 	if reflect.DeepEqual(src, dst) {
 		t.Error("destination should not match source")
+	}
+}
+
+// ss returns a sorted slice of byte slices.
+func ss(s ...string) [][]byte {
+	r := make([][]byte, len(s))
+	for i := range s {
+		r[i] = []byte(s[i])
+	}
+	bytesutil.Sort(r)
+	return r
+}
+
+func TestCompareSlice(t *testing.T) {
+	name := func(a, b [][]byte, exp int) string {
+		return fmt.Sprintf("%s <=> %s is %d", bytes.Join(a, nil), bytes.Join(b, nil), exp)
+	}
+	tests := []struct {
+		a, b [][]byte
+		exp  int
+	}{
+		{
+			a:   ss("aaa", "bbb", "ccc"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: 0,
+		},
+
+		{
+			a:   ss("aaa", "bbb", "ccc", "ddd"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: 1,
+		},
+
+		{
+			a:   ss("aaa", "bbb"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: -1,
+		},
+
+		{
+			a:   ss("aaa", "bbbb"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: 1,
+		},
+
+		{
+			a:   ss("aaa", "ccc"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: 1,
+		},
+
+		{
+			a:   ss("aaa", "aaa"),
+			b:   ss("aaa", "bbb", "ccc"),
+			exp: -1,
+		},
+	}
+	for _, test := range tests {
+		t.Run(name(test.a, test.b, test.exp), func(t *testing.T) {
+			if got := CompareSlice(test.a, test.b); got != test.exp {
+				t.Errorf("unexpected result, -got/+exp\n%s", cmp.Diff(got, test.exp))
+			}
+		})
 	}
 }
